+++ carla_ros_bridge/src/carla_ros_bridge/walker.py
@@ -11,14 +11,13 @@ Classes to handle Carla pedestrians
 """
 import rospy
 from derived_object_msgs.msg import Object
-from shape_msgs.msg import SolidPrimitive
 
-from carla_ros_bridge.actor import Actor
+from carla_ros_bridge.traffic_participant import TrafficParticipant
 from carla_msgs.msg import CarlaWalkerControl
 from carla import WalkerControl
 
 
-class Walker(Actor):
+class Walker(TrafficParticipant):
 
     """
     Actor implementation details for pedestrians
@@ -45,9 +44,6 @@ class Walker(Actor):
                                      communication=communication,
                                      prefix=prefix)
 
-        self.classification = Object.CLASSIFICATION_PEDESTRIAN
-        self.classification_age = 0
-
         self.control_subscriber = rospy.Subscriber(
             self.get_topic_prefix() + "/walker_control_cmd",
             CarlaWalkerControl, self.control_command_updated)
@@ -71,52 +67,9 @@ class Walker(Actor):
         walker_control.jump = ros_walker_control.jump
         self.carla_actor.apply_control(walker_control)
 
-    def update(self, frame, timestamp):
+    def get_classification(self):
         """
-        Function (override) to update this object.
-
-        On update walkers send:
-        - tf global frame
-        - object message
-        - marker message
-
+        Function (override) to get classification
         :return:
         """
-        self.classification_age += 1
-        self.publish_transform(self.get_ros_transform())
-        self.publish_marker()
-        super(Walker, self).update(frame, timestamp)
-
-    def get_object_info(self):
-        """
-        Function to send object messages of this walker
-
-        A derived_object_msgs.msg.Object is prepared to be published via '/carla/objects'
-
-        :return:
-        """
-        walker_object = Object(header=self.get_msg_header("map"))
-        # ID
-        walker_object.id = self.get_id()
-        # Pose
-        walker_object.pose = self.get_current_ros_pose()
-        # Twist
-        walker_object.twist = self.get_current_ros_twist()
-        # Acceleration
-        walker_object.accel = self.get_current_ros_accel()
-        # Shape
-        walker_object.shape.type = SolidPrimitive.BOX
-        walker_object.shape.dimensions.extend([
-            self.carla_actor.bounding_box.extent.x * 2.0,
-            self.carla_actor.bounding_box.extent.y * 2.0,
-            self.carla_actor.bounding_box.extent.z * 2.0])
-
-        # Classification if available in attributes
-        if self.classification != Object.CLASSIFICATION_UNKNOWN:
-            walker_object.object_classified = True
-            walker_object.classification = self.classification
-            walker_object.classification_certainty = 1.0
-            self.classification_age += 1
-            walker_object.classification_age = self.classification_age
-
-        return walker_object
+        return Object.CLASSIFICATION_PEDESTRIAN